# backend/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Stellar Payment API",
    "version": "0.1.0",
    "description": "API for creating and verifying Stellar network payments"
  },
  "servers": [
    {
      "url": "http://localhost:4000"
    }
  ],
  "paths": {
...
    "/api/auth/challenge": {
      "post": {
        "summary": "Generate a SEP-0010 challenge transaction",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "account"
                ],
                "properties": {
                  "account": {
                    "type": "string",
                    "description": "Stellar public key (G...)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Challenge transaction",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "transaction": {
                      "type": "string",
                      "description": "Base64-encoded challenge XDR"
                    },
                    "network_passphrase": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          }
        }
      }
    },
    "/api/auth/verify": {
      "post": {
        "summary": "Verify a signed SEP-0010 challenge and issue session token",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "transaction"
                ],
                "properties": {
                  "transaction": {
                    "type": "string",
                    "description": "Signed challenge transaction XDR"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentication successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "description": "Session JWT token"
                    },
                    "merchant": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed"
          }
        }
      }
    },
...
    "/api/rotate-key": {
      "post": {
        "summary": "Rotate the authenticated merchant's API key",
        "tags": [
          "Merchants"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "New API key issued; the old key is immediately invalidated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "api_key": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid x-api-key header"
          }
        }
      }
    },
...
    "/api/create-payment": {
      "post": {
        "summary": "Create a new payment session request",
        "tags": [
          "Payments"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "schema": {
              "type": "string"
            },
            "description": "Optional unique key for idempotent requests. Use UUID or request ID. Responses are cached for 24 hours."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "amount",
                  "asset",
                  "recipient"
                ],
                "properties": {
                  "amount": {
                    "type": "number",
                    "description": "Payment amount (must be positive and at least 0.01 XLM for native payments)"
                  },
                  "asset": {
                    "type": "string",
                    "description": "Asset code (e.g. XLM, USDC)"
                  },
                  "asset_issuer": {
                    "type": "string",
                    "description": "Asset issuer (required for non-native assets)"
                  },
                  "recipient": {
                    "type": "string",
                    "description": "Stellar address of the recipient"
                  },
                  "merchant_id": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "memo": {
                    "type": "string"
                  },
                  "memo_type": {
                    "type": "string",
                    "enum": [
                      "text",
                      "id",
                      "hash",
                      "return"
                    ]
                  },
                  "webhook_url": {
                    "type": "string"
                  },
                  "branding_overrides": {
                    "type": "object",
                    "properties": {
                      "primary_color": {
                        "type": "string",
                        "example": "#5ef2c0"
                      },
                      "secondary_color": {
                        "type": "string",
                        "example": "#b8ffe2"
                      },
                      "background_color": {
                        "type": "string",
                        "example": "#050608"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Duplicate request — cached response returned from idempotency key",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "payment_id": {
                      "type": "string"
                    },
                    "payment_link": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "201": {
            "description": "Payment created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "payment_id": {
                      "type": "string"
                    },
                    "payment_link": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    },
                    "branding_config": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error or invalid Idempotency-Key"
          },
          "429": {
            "description": "Too many requests"
          }
        }
      }
    },
...
    "/api/payments": {
      "get": {
        "summary": "Get paginated list of payments for the authenticated merchant",
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number (1-indexed)"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Number of results per page (max 100)"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated payments",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "payments": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "total_count": {
                      "type": "integer"
                    },
                    "total_pages": {
                      "type": "integer"
                    },
                    "page": {
                      "type": "integer"
                    },
                    "limit": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key"
          }
        }
      }
    },
...
  },
...
}